Consolidate environment requirements-related `.go` files
