+++ internal/app/forklift/env-models.go
@@ -35,51 +35,6 @@ type EnvSpec struct {
 	Description string `yaml:"description,omitempty"`
 }
 
-// Repo Requirements
-
-const (
-	// RepoReqsDirName is the directory in a Forklift environment which contains Pallet
-	// repository requirement configurations.
-	// TODO: move repositories to requirements/repositories, to allow for a future
-	// requirements/environments subdirectory.
-	RepoReqsDirName = "repositories"
-)
-
-// A FSRepoReq is a Pallet repository requirement stored at the root of a [fs.FS]
-// filesystem.
-type FSRepoReq struct {
-	// RepoReq is the Pallet repository requirement at the root of the filesystem.
-	RepoReq
-	// FS is a filesystem which contains the repository requirement's contents.
-	FS pallets.PathedFS
-}
-
-// A RepoReq is a requirement for a specific Pallet repository at a specific version.
-type RepoReq struct {
-	// VCSRepoPath is the VCS repository path of the required repository.
-	VCSRepoPath string
-	// RepoSubdir is the Pallet repository subdirectory of the required repository.
-	RepoSubdir string
-	// VersionLock specifies the version of the required repository.
-	VersionLock VersionLock
-}
-
-// Package Requirements
-
-// A PkgReq is a requirement for a Pallet package at a specific version.
-type PkgReq struct {
-	// PkgSubdir is the Pallet package subdirectory of the repository which should provide the
-	// required package.
-	PkgSubdir string
-	// Repo is the Pallet repository which should provide the required package.
-	Repo RepoReq
-}
-
-// PkgReqLoader is a source of Pallet package requirements.
-type PkgReqLoader interface {
-	LoadPkgReq(pkgPath string) (PkgReq, error)
-}
-
 // Deployment Configurations
 
 const (
@@ -116,4 +71,47 @@ type DeplConfig struct {
 	// Features is a list of features from the Pallet package which should be enabled in the
 	// deployment.
 	Features []string `yaml:"features,omitempty"`
+}
+
+// Requirements
+
+// A PkgReq is a requirement for a Pallet package at a specific version.
+type PkgReq struct {
+	// PkgSubdir is the Pallet package subdirectory of the repository which should provide the
+	// required package.
+	PkgSubdir string
+	// Repo is the Pallet repository which should provide the required package.
+	Repo RepoReq
+}
+
+// PkgReqLoader is a source of Pallet package requirements.
+type PkgReqLoader interface {
+	LoadPkgReq(pkgPath string) (PkgReq, error)
+}
+
+const (
+	// RepoReqsDirName is the directory in a Forklift environment which contains Pallet
+	// repository requirement configurations.
+	// TODO: move repositories to requirements/repositories, to allow for a future
+	// requirements/environments subdirectory.
+	RepoReqsDirName = "repositories"
+)
+
+// A FSRepoReq is a Pallet repository requirement stored at the root of a [fs.FS]
+// filesystem.
+type FSRepoReq struct {
+	// RepoReq is the Pallet repository requirement at the root of the filesystem.
+	RepoReq
+	// FS is a filesystem which contains the repository requirement's contents.
+	FS pallets.PathedFS
+}
+
+// A RepoReq is a requirement for a specific Pallet repository at a specific version.
+type RepoReq struct {
+	// VCSRepoPath is the VCS repository path of the required repository.
+	VCSRepoPath string
+	// RepoSubdir is the Pallet repository subdirectory of the required repository.
+	RepoSubdir string
+	// VersionLock specifies the version of the required repository.
+	VersionLock VersionLock
 }
